some refactors

+++ src/app/app.cpp
@@ -7,7 +7,7 @@ App::App() {
 }
 
 int main() {
-    auto seed = std::chrono::system_clock::now().time_since_epoch().count();
+    auto seed = static_cast<unsigned int>(std::chrono::system_clock::now().time_since_epoch().count());
     std::default_random_engine gen2(seed);
 
     App* app = new App;
